增加客户客户组员工组删除弹窗

+++ src/views/staff/staffgrouplist.vue
@@ -136,23 +136,29 @@ export default {
     },
     // 删除
     deleteData () {
-      console.log(this.deleteNum, '2222')
-      this.$request.deleteGroup({
-        DELETE: this.deleteNum
-      }).then(res => {
-        if (res.ret) {
-          this.$notify({
-            title: '操作成功',
-            message: res.msg,
-            type: 'success'
-          })
-          this.getList()
-        } else {
-          this.$message({
-            message: res.msg,
-            type: 'error'
-          })
-        }
+      this.$confirm(`是否确认删除？`, '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        console.log(this.deleteNum, '2222')
+        this.$request.deleteGroup({
+          DELETE: this.deleteNum
+        }).then(res => {
+          if (res.ret) {
+            this.$notify({
+              title: '操作成功',
+              message: res.msg,
+              type: 'success'
+            })
+            this.getList()
+          } else {
+            this.$message({
+              message: res.msg,
+              type: 'error'
+            })
+          }
+        })
       })
     },
     selectionChange (selection) {
