+++ src/views/staff/stafflist.vue
@@ -168,22 +168,28 @@ export default {
     },
     // 删除
     deleteData () {
-      this.$request.deleteVip({
-        DELETE: this.deleteNum
-      }).then(res => {
-        if (res.ret) {
-          this.$notify({
-            title: '操作成功',
-            message: res.msg,
-            type: 'success'
-          })
-          this.getList()
-        } else {
-          this.$message({
-            message: res.msg,
-            type: 'error'
-          })
-        }
+      this.$confirm(`是否确认删除？`, '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        this.$request.deleteVip({
+          DELETE: this.deleteNum
+        }).then(res => {
+          if (res.ret) {
+            this.$notify({
+              title: '操作成功',
+              message: res.msg,
+              type: 'success'
+            })
+            this.getList()
+          } else {
+            this.$message({
+              message: res.msg,
+              type: 'error'
+            })
+          }
+        })
       })
     },
     // 禁止/允许登录
